pyproject.toml
[project]
name = "mcp-ad-network"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.59.0",
    "claudesync>=0.7.3",
    "gradio>=5.3.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.2",
    "openai>=1.97.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlite-vec>=0.1.6",
]
